+++ src/auth0__jwt__expiration__is_valid_/index.d.ts
@@ -1,0 +1,1 @@
+export declare function auth0__jwt__expiration__is_valid_(jwt_token:string):boolean