+++ src/auth0__jwt__expiration__is_valid_/index.js
@@ -1,0 +1,7 @@
+import { jwt__expiration__is_valid_ } from '@ctx-core/jwt'
+/**
+ * @param {string}jwt_token
+ */
+export function auth0__jwt__expiration__is_valid_(jwt_token) {
+	return jwt__expiration__is_valid_(jwt_token)
+}