+++ src/auth0__token__is_valid_/index.d.ts
@@ -1,0 +1,2 @@
+import type { auth0__token_T } from '../auth0__token__'
+export declare function auth0__token__is_valid_(auth0_token:auth0__token_T):boolean